+++ hahomematic/model/generic/switch.py
@@ -18,6 +18,8 @@ class DpSwitch(GenericDataPoint[bool | None, bool]):
     This is a default data point that gets automatically generated.
     """
 
+    __slots__ = ()
+
     _category = DataPointCategory.SWITCH
 
     @state_property