+++ changelog.md
@@ -1,6 +1,11 @@
+# Version 2025.8.1 (2025-08-05)
+
+- Cleanup slots
+- Move timer support to BaseDataPoint
+
 # Version 2025.8.0 (2025-08-03)
 
-- Use **slots**
+- Use slots
 
 # Version 2025.7.7 (2025-07-26)